% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_children.R
\name{addChildren}
\alias{addChildren}
\title{Add child nodes to the passed parent node}
\usage{
addChildren(node, avps)
}
\arguments{
\item{node}{An \code{xml2} node}

\item{avps}{A named list}
}
\value{
An \code{xml2} node with the children appended
}
\description{
Add child nodes to the passed parent node
}
\examples{
library("xml2")
parent    <- read_xml("<body/>")
add_nodes <- list(foo = "bar", fizz = "buzz")
parent    <- addressr::addChildren(parent, add_nodes)
}
