% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{address_abbreviations}
\alias{address_abbreviations}
\title{Address abbreviations}
\format{
\subsection{\code{address_abbreviations}}{

A data frame with 542 rows and 3 columns:
\describe{
\item{short}{Abbreviation}
\item{long}{Full word}
\item{type}{Category of the abbreviation}
}
}
}
\usage{
address_abbreviations
}
\description{
Commonly used abbreviations in addresses.
}
\keyword{datasets}
