% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{address_corrections}
\alias{address_corrections}
\title{Address corrections}
\format{
\subsection{\code{address_corrections}}{

A dataframe with 17 rows and 4 columns:
\describe{
\item{input}{The original string found in the data}
\item{output}{The string to output}
\item{city}{The city/place for the correction}
\item{state}{The state for the correction}
}
}
}
\usage{
address_corrections
}
\description{
Corrections to address errors
}
\keyword{datasets}
