% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addressr-package.R
\docType{package}
\name{addressr-package}
\alias{addressr}
\alias{addressr-package}
\title{addressr: Clean Addresses}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Sarah Johnson \email{sarah.johnson@princeton.edu}

}
\keyword{internal}
