% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{check_pattern}
\alias{check_pattern}
\title{Check the address pattern}
\usage{
check_pattern(pattern)
}
\arguments{
\item{pattern}{The string to be extracted and removed.}
}
\value{
A string to use in pattern matching.
}
\description{
Check the address pattern
}
