% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkers.R
\name{check_unit}
\alias{check_unit}
\title{Check the unit}
\usage{
check_unit(
  .data,
  unit,
  unit_type,
  street_number,
  street_suffix,
  building,
  addressr_id
)
}
\arguments{
\item{.data}{A data frame, data frame extension (e.g. a tibble), or a lazy data frame (e.g. from dbplyr or dtplyr).}

\item{unit}{The unit number or letter}

\item{unit_type}{The type of unit (apt, #, etc)}

\item{street_number}{The street number (<123> N Main St)}

\item{street_suffix}{The street suffix (123 N Main \if{html}{\out{<St>}})}

\item{building}{The building number or letter}

\item{addressr_id}{Unique row_id
\enumerate{
\item remove all non-word characters
\item extract numerics, compare with street number
\item if a unit contains text that is more than 2 characters, put it in "extra_unit" to fix later
\item extract words, compare with street suffix & building
\item combine unit_number and unit_text
}}
}
\description{
Check the unit
}
