% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_address.R
\name{cleanAddress}
\alias{cleanAddress}
\title{Clean a raw address}
\usage{
cleanAddress(userid, data, address_column = "Address", max_tries = 10L)
}
\arguments{
\item{userid}{The userid given to you by USPS, see the Details for more}

\item{data}{A data.frame or \code{\link[data.table]{data.table}} with at least
a column containing raw addresses}

\item{address_column}{The name of the column containing the addresses}

\item{max_tries}{The maximum iterations to attempt when geocoding, see the Details for more}
}
\description{
Clean a raw address
}
\details{
The \code{\link[ggmap]{geocode}} function seems to throw spurious `OVER_QUERY_LIMIT`
    errors quite frequently so the `cleanAddress` function has some basic handling
    of these errors, allowing a certain number of retries for those addresses that did
    not get geocoded. When an address is successfully geocoded it will not be
    retried.

    If you do not have a `userid` you should register at \url{https://www.usps.com/business/web-tools-apis}
    before trying to use this function.
}
\examples{
\dontrun{
}
}
\seealso{
\code{\link{validateAddress}}
}
