% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usps.R
\name{compose_xml}
\alias{compose_xml}
\title{Recompose address to follow USPS API specs}
\usage{
compose_xml(usr, street, city, state, zip = NULL)
}
\arguments{
\item{zip}{}
}
\value{

}
\description{
Recompose address to follow USPS API specs
}
\examples{
compose_xml(USPS_USER_KEY, "123 Main Street", "Chicago", "IL", "60421")
}
