% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usps.R
\name{executeCall}
\alias{executeCall}
\title{Execute API call(s) to USPS}
\usage{
executeCall(req)
}
\arguments{
\item{req}{API request recomposed from calling \code{compose_xml}}
}
\value{
data frame with cols: Street, City, State, Zip, Zip4
}
\description{
Execute API call(s) to USPS
}
