% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_remove_squish.R
\name{extract_remove_squish_db}
\alias{extract_remove_squish_db}
\title{Extract, remove, and squish a string in a dataframe}
\usage{
extract_remove_squish_db(.data, original_column, new_column, pattern)
}
\arguments{
\item{.data}{A data frame, data frame extension (e.g. a tibble), or a lazy data frame (e.g. from dbplyr or dtplyr).}

\item{original_column}{The column from which the string should be extracted, then removed, then squished to remove extra whitespace.}

\item{new_column}{The name of the column to output the extracted strings.}

\item{pattern}{The string to be extracted and removed.}
}
\value{
An object of the same type as .data, with the following properties:
\itemize{
\item A modified original column, from which the pattern was removed and whitespace was trimmed.
\item A new column containing the extracted string.
}
}
\description{
Extract, remove, and squish a string in a dataframe
}
