% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{str_collapse_bound}
\alias{str_collapse_bound}
\title{Flatten and bound a vector for use in regex}
\usage{
str_collapse_bound(vector)
}
\arguments{
\item{vector}{A vector or something that can be coerced to one.}
}
\value{
A flattened string bound with "\\b(string)\\b)
}
\description{
Flatten and bound a vector for use in regex
}
