% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{str_replace_names}
\alias{str_replace_names}
\title{Replace a string with another string}
\usage{
str_replace_names(string, input, output)
}
\arguments{
\item{string}{A string (or vector, column, etc) to be altered}

\item{input}{Pattern to find in the existing string}

\item{output}{Replacement for the matched pattern}
}
\value{
A modified version of the original string
}
\description{
Replace a string with another string
}
