% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{switch_abbreviation}
\alias{switch_abbreviation}
\title{Replace a string using the address_abbreviations table}
\usage{
switch_abbreviation(string, type, method = "long-to-short")
}
\arguments{
\item{string}{A string (or vector, column, etc) to be altered}

\item{type}{The category of strings to filter for from \code{address_abbreviations$type}}

\item{method}{\code{"long-to-short"} (the default) to abbreviate the string, or \code{"short-to-long"} to un-abbreviate.}
}
\value{
A modified version of the original string
}
\description{
Replace a string using the address_abbreviations table
}
