% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{table_reference}
\alias{table_reference}
\title{Table reference}
\format{
\subsection{\code{table_reference}}{

A dataframe with 78 rows and 5 columns
\describe{
\item{table_name}{The name of the table}
\item{original_column}{The name of the column as originally found in the data}
\item{rename_column}{A column name to replace the original name}
\item{data_category}{The category of the data column}
\item{data_type}{The data type of the column}
}
}
}
\usage{
table_reference
}
\description{
A reference for commonly used tables
}
\keyword{datasets}
