% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{to_building_address}
\alias{to_building_address}
\title{Regex standardization to convert unit level address to building level}
\usage{
to_building_address(x, removeSuffix = TRUE)
}
\arguments{
\item{x}{the raw address as the input}

\item{removeSuffix}{if TRUE(default), remove the street suffix like St, Ave, Blvd, etc}
}
\value{
cleansed address at building level
}
\description{
Regex standardization to convert unit level address to building level
}
\examples{
to_building_address("123 Main Street, Unit #1", removeSuffix=TRUE)
}
