% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_address.R
\name{clean_address}
\alias{clean_address}
\title{Clean Addresses}
\usage{
clean_address(
  .data,
  input_column,
  dataset = "default",
  separate_street_range = TRUE
)
}
\arguments{
\item{.data}{A data frame, data frame extension (e.g. a tibble), or a lazy data frame (e.g. from dbplyr or dtplyr).}

\item{input_column}{The column from which the string should be extracted, then removed, then squished to remove extra whitespace.}

\item{dataset}{The dataset to clean. Either "default", "quick", or "default_db"}

\item{separate_street_range}{Should street numbers with a range be pivoted into individual rows?}
}
\value{
An object of the same type as .data, with the following properties:
\itemize{
\item A modified original column, from which the pattern was removed and whitespace was trimmed.
\item New column containing the extracted strings.
}
}
\description{
Clean Addresses
}
