% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_addresses}
\alias{test_addresses}
\title{Test address dataset}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 208 rows and 18 columns.
}
\usage{
test_addresses
}
\description{
Dataset to show how addresses are formatted with addressr.
}
\keyword{datasets}
